week2/code/requirements.txt
# Week 2 基线依赖
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=8.0.0
